Change API endpoint and hide pop up after close
